// demo/merchant/src/models/types.ts
  export interface Item {
    /** Id of a piece of merchandise that can be purchased */
    id: string;
    /** Quantity of the item for fulfillment (minimum 1) */
    quantity: number;
  }
// ...
  export interface LineItem {
    /** Id of the line item (different from the item id) */
    id: string;
    /** Item that is represented by the line item */
    item: Item;
    /** Integer representing item base amount before adjustments */
    base_amount: number;
    /** Integer representing any discount applied to the item */
    discount: number;
    /** Integer representing amount after all adjustments */
    subtotal: number;
    /** Integer representing tax amount */
    tax: number;
    /** Integer representing total amount */
    total: number;
  }
  
  /**
   * Total types
   */
  export type TotalType =
    | 'items_base_amount'
    | 'items_discount'
    | 'subtotal'
    | 'discount'
    | 'fulfillment'
    | 'tax'
    | 'fee'
    | 'total';
  
  /**
   * Total - Price breakdown component
   */
  export interface Total {
    /** String value representing the type of total */
    type: TotalType;
    /** The text displayed to the customer for this total */
    display_text: string;
    /** Integer representing total amount in minor units */
    amount: number;
  }